Add UUID identifier in new version format
The requirement for contiguous, persistent integer id's
hinders the composability of multi-source migrations. This
commit introduces a UUID identifier for migrations, with
ordering now being as given, instead of the integer based.

This version is fully complatible with API and database
changelogs for older versions. Existing projects that wish
to move to the UUID identifier, will need to perform a
stepwise migration from format 1 over format 2 to 3, to
associate the new identifier with existing changelog entries
in running environments.

It is supported to migrate all the way *back* to version 1,
from version 3 (as long as it's over version 2), so rollback
to a previous version should generally be possible, depending
on the compatibility of the schema changes of the new version,
with the application logic of the previous version.

Binary compatibility with previous release has been preserved.
on the call interfacer, but not with the values contained of
the response type AppliedMigrations.